fixed the publication form to make it look shorter

+++ app/Http/Controllers/Admin/CommsOfPracticeController.php
@@ -81,7 +81,9 @@ class CommsOfPracticeController extends Controller
 
 
     public function destroy(Request $request){
-
+        if (!auth()->user() || !auth()->user()->can('delete_meta_data')) {
+            return response(['status'=>'failure','message'=>'Unauthorized'], 403);
+        }
         $deleted  = $this->commsOfPracticeRepository->delete($request->id);
 
         if($deleted):
